+++ lizard_wbconfiguration/api/views.py
@@ -168,7 +168,8 @@ class WaterBalanceDBF(View):
         success = False
 
         mapping = WBConfigurationDBFMapping.objects.filter(
-            model_name__iexact=model_name)
+            model_name__iexact=model_name).order_by('index')
+        print mapping
 
         try:
             logger.debug("Create en open dbf file.")